+++ App Web/RED/src/main/java/it/red/institution/Institution.java
@@ -2,23 +2,23 @@ package it.red.institution;
 
 
 public class Institution {
-	private int id_institution;
+	private long id_institution;
 	private String name;
 
 	public Institution() {
 		
 	}
 	
-	public Institution(int id_institution, String name) {
+	public Institution(long id_institution, String name) {
 		this.id_institution = id_institution;
 		this.name = name;
 	}
 
-	public int getId_institution() {
+	public long getId_institution() {
 		return id_institution;
 	}
 
-	public void setId_institution(int id_institution) {
+	public void setId_institution(long id_institution) {
 		this.id_institution = id_institution;
 	}
