+++ App Web/RED/src/main/java/it/red/lesson/Lesson.java
@@ -1,15 +1,15 @@
 package it.red.lesson;
 
 public class Lesson {
-	private int id_lesson;
-	private int agenda_id_agenda;
+	private long id_lesson;
+	private long agenda_id_agenda;
 	private boolean presence;
-	private int grade;
-	private int student_id_student;
+	private long grade;
+	private long student_id_student;
 	
 	public Lesson() {}
 	
-	public Lesson(int id_lesson, int agenda_id_agenda, boolean presence, int grade, int student_id_student) {
+	public Lesson(long id_lesson, long agenda_id_agenda, boolean presence, long grade, long student_id_student) {
 		this.id_lesson = id_lesson;
 		this.agenda_id_agenda = agenda_id_agenda;
 		this.presence = presence;
@@ -17,18 +17,18 @@ public class Lesson {
 		this.student_id_student = student_id_student;
 	}
 	
-	public int getId_lesson() {
+	public long getId_lesson() {
 		return id_lesson;
 	}
 	
-	public void setId_lesson(int id_lesson) {
+	public void setId_lesson(long id_lesson) {
 		this.id_lesson = id_lesson;
 	}
 	
-	public int getAgenda_id_agenda() {
+	public long getAgenda_id_agenda() {
 		return agenda_id_agenda;
 	}
-	public void setAgenda_id_agenda(int agenda_id_agenda) {
+	public void setAgenda_id_agenda(long agenda_id_agenda) {
 		this.agenda_id_agenda = agenda_id_agenda;
 	}
 	
@@ -39,19 +39,19 @@ public class Lesson {
 		this.presence = presence;
 	}
 	
-	public int getGrade() {
+	public long getGrade() {
 		return grade;
 	}
 	
-	public void setGrade(int grade) {
+	public void setGrade(long grade) {
 		this.grade = grade;
 	}
 	
-	public int getStudent_id_student() {
+	public long getStudent_id_student() {
 		return student_id_student;
 	}
 	
-	public void setStudent_id_student(int student_id_student) {
+	public void setStudent_id_student(long student_id_student) {
 		this.student_id_student = student_id_student;
 	}
